==> notebooks/04_model_advanced.ipynb <==
## 1. Import Libraries

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix, f1_score
from sklearn.impute import SimpleImputer

from imblearn.over_sampling import SMOTE
from xgboost import XGBClassifier

import warnings
warnings.filterwarnings("ignore")


## 2. Load Advanced Engineered Dataset

df = pd.read_csv("../data/processed/credit_final_engineered.csv")
df.head()
df.columns

## 3. Prepare X and y

X = df.drop("default payment next month", axis=1)
y = df["default payment next month"]


## 4. Train–Test Split

X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.2,
    random_state=42,
    stratify=y
)


## 5. Replace Inf → NaN → Impute Missing Values

X_train = X_train.replace([np.inf, -np.inf], np.nan)
X_test = X_test.replace([np.inf, -np.inf], np.nan)

imputer = SimpleImputer(strategy="median")
X_train = imputer.fit_transform(X_train)
X_test = imputer.transform(X_test)


## 6. Apply SMOTE (Handle Imbalance)

smote = SMOTE(random_state=42)
X_train_sm, y_train_sm = smote.fit_resample(X_train, y_train)


## 7. Train Final Tuned XGBoost Model

final_xgb = XGBClassifier(
    n_estimators=450,
    max_depth=5,
    learning_rate=0.07,
    subsample=0.8,
    colsample_bytree=0.8,
    min_child_weight=1,
    gamma=0.1,
    eval_metric='logloss',
    random_state=42
)

final_xgb.fit(X_train_sm, y_train_sm)


## 8. Predict Using Probabilities

y_prob = final_xgb.predict_proba(X_test)[:, 1]


## 9. Threshold Optimization

thresholds = np.arange(0.1, 0.9, 0.01)

best_f1 = 0
best_threshold = 0.5

for t in thresholds:
    preds = (y_prob > t).astype(int)
    score = f1_score(y_test, preds)
    if score > best_f1:
        best_f1 = score
        best_threshold = t

print("Best Threshold:", best_threshold)
print("Best F1:", best_f1)


## 10. Final Predictions Using Best Threshold

y_final = (y_prob > best_threshold).astype(int)

print("Final Accuracy:", accuracy_score(y_test, y_final))
print("\nClassification Report:\n", classification_report(y_test, y_final))
print("\nConfusion Matrix:\n", confusion_matrix(y_test, y_final))


## 11. Save Final Model

final_xgb.save_model("../models/final_xgb_model.json")
print("Model saved!")
